refactor(cookie-consent): address review feedback (#2340)
- rename cookie-consent.registry.ts to cookie-consent.consts.ts for codebase consistency
- derive Copilot host selector from exported COPILOT_SCRIPT_ID instead of hardcoding
- use filigran-ui Button in CookieSettingsLink and CookieConsentPreferences
