+++ apps/frontend/src/components/cookie-consent/cookie-consent.consts.ts
