+++ apps/frontend/src/components/cookie-consent/cookie-consent.server.ts
@@ -1,4 +1,4 @@
-import { CONSENT_COOKIE_NAME } from '@/components/cookie-consent/cookie-consent.registry';
+import { CONSENT_COOKIE_NAME } from '@/components/cookie-consent/cookie-consent.consts';
 import { type StoredConsent } from '@/components/cookie-consent/cookie-consent.types';
 import { parseStoredConsent } from '@/components/cookie-consent/cookie-consent.utils';
 import { cookies } from 'next/headers';
